Added bounds checking and snapping to positions for selections
git-svn-id: http://svn.powertabs.org/svn/ptedev/trunk@261 5b1460fe-822e-4aa1-ac1b-8a6412c92c8c

+++ source/painters/staffpainter.h
@@ -29,8 +29,8 @@ protected:
     StaffData staffInfo;
     QPen pen;
 
-    double selectionStart;
-    double selectionEnd;
+    int selectionStart;
+    int selectionEnd;
 
     void init();
 };
